Commit final

+++ backend/controllers/tareasController.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/tareaModel')
 
 const getTareas = asyncHandler(async (req, res) => {
 
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user._id })
 
     res.status(200).json(tareas)
 })
@@ -15,7 +15,8 @@ const createTarea = asyncHandler(async (req, res) => {
     }
 
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto,
+        user: req.user._id
     })
 
     res.status(201).json(tarea)
@@ -25,15 +26,20 @@ const updateTarea = asyncHandler(async (req, res) => {
 
     const tarea = await Tarea.findById(req.params.id)
 
+    //verificamos que la tarea existe
     if (!tarea) {
         res.status(400)
         throw new Error("Tarea no encontrada")
     }
 
-    const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-
-
-    res.status(200).json(updatedTarea)
+    //verificamos que la tarea le pertenece al usuario del token proporcionado
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error("Acceso no autorizado")
+    } else {
+        const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        res.status(200).json(updatedTarea)
+    }
 })
 
 const deleteTarea = asyncHandler(async (req, res) => {
@@ -45,10 +51,15 @@ const deleteTarea = asyncHandler(async (req, res) => {
         throw new Error("Tarea no encontrada")
     }
 
-    await Tarea.deleteOne(tarea)
-    //2. const deletedTarea = await Tarea.findByIdAndDelete(req.params.id)
-
-    res.status(200).json({ id: req.params.id })
+    //verificamos que la tarea le pertenece al usuario del token proporcionado
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error("Acceso no autorizado")
+    } else {
+        await Tarea.deleteOne(tarea)
+        //2. const deletedTarea = await Tarea.findByIdAndDelete(req.params.id)
+        res.status(200).json({ id: req.params.id })
+    }
 })
 
 module.exports = {
